requirements.txt
Flask==1.0.2
gevent==1.3.6
redis==2.10.6
lxml==4.2.5
requests==2.20.1
PyMySQL==0.9.3
kafka-python==1.4.4